+++ src/presenters/pages/collection.js
@@ -18,6 +18,7 @@ import CollectionNameInput from 'Components/fields/collection-name-input';
 import DataLoader from 'Components/data-loader';
 import MoreCollectionsContainer from 'Components/collections-list/more-collections';
 import ReportButton from 'Components/report-abuse-pop';
+import EditCollectionColor from 'Components/collection/edit-collection-color-pop';
 import { AnalyticsContext } from 'State/segment-analytics';
 import { useCurrentUser } from 'State/current-user';
 import { getSingleItem, getAllPages } from 'Shared/api';
@@ -25,7 +26,6 @@ import { getSingleItem, getAllPages } from 'Shared/api';
 import Layout from '../layout';
 import AuthDescription from '../includes/auth-description';
 import CollectionEditor from '../collection-editor';
-import EditCollectionColor from '../pop-overs/edit-collection-color-pop';
 import AddCollectionProject from '../pop-overs/add-collection-project-pop';
 import CollectionAvatar from '../includes/collection-avatar';
